final_project/capstone/requirements.txt
Django
django-filter
django-multiselectfield
Pillow